fix(LibraryDetail): extract categoryChecker function to library helpers
